+++ strategies/beaconblockproposal/best/beaconblockproposal.go
@@ -1,4 +1,4 @@
-// Copyright © 2020 Attestant Limited.
+// Copyright © 2020 - 2022 Attestant Limited.
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
@@ -93,11 +93,11 @@ func (s *Service) BeaconBlockProposal(ctx context.Context, slot phase0.Slot, ran
 				if err != nil {
 					log.Error().Str("version", proposal.Version.String()).Err(err).Msg("Failed to obtain proposal slot for parent block")
 				}
-				parentSlot = slot - 1
+				parentSlot = slot
 			}
 
 			mu.Lock()
-			score := scoreBeaconBlockProposal(ctx, name, parentSlot, proposal)
+			score := s.scoreBeaconBlockProposal(ctx, name, parentSlot, proposal)
 			if score > bestScore || bestProposal == nil {
 				bestScore = score
 				bestProposal = proposal