+++ strategies/beaconblockproposal/best/service.go
@@ -1,4 +1,4 @@
-// Copyright © 2020 Attestant Limited.
+// Copyright © 2020, 2022 Attestant Limited.
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
@@ -15,11 +15,15 @@ package best
 
 import (
 	"context"
+	"sync"
 	"time"
 
 	eth2client "github.com/attestantio/go-eth2-client"
+	"github.com/attestantio/go-eth2-client/spec/phase0"
+	"github.com/attestantio/vouch/services/chaintime"
 	"github.com/attestantio/vouch/services/metrics"
 	"github.com/pkg/errors"
+	"github.com/prysmaticlabs/go-bitfield"
 	"github.com/rs/zerolog"
 	zerologger "github.com/rs/zerolog/log"
 )
@@ -28,16 +32,37 @@ import (
 type Service struct {
 	clientMonitor                metrics.ClientMonitor
 	processConcurrency           int64
+	chainTime                    chaintime.Service
 	beaconBlockProposalProviders map[string]eth2client.BeaconBlockProposalProvider
 	signedBeaconBlockProvider    eth2client.SignedBeaconBlockProvider
 	timeout                      time.Duration
+
+	// Spec values for scoring proposals.
+	slotsPerEpoch      uint64
+	timelySourceWeight uint64
+	timelyTargetWeight uint64
+	timelyHeadWeight   uint64
+	syncRewardWeight   uint64
+	proposerWeight     uint64
+	weightDenominator  uint64
+
+	priorBlocks   map[phase0.Root]*priorBlockVotes
+	priorBlocksMu sync.RWMutex
+}
+
+type priorBlockVotes struct {
+	root   phase0.Root
+	parent phase0.Root
+	slot   phase0.Slot
+	// votes is a map of attestation slot -> committee index -> votes
+	votes map[phase0.Slot]map[phase0.CommitteeIndex]bitfield.Bitlist
 }
 
 // module-wide log.
 var log zerolog.Logger
 
 // New creates a new beacon block propsal strategy.
-func New(_ context.Context, params ...Parameter) (*Service, error) {
+func New(ctx context.Context, params ...Parameter) (*Service, error) {
 	parameters, err := parseAndCheckParameters(params...)
 	if err != nil {
 		return nil, errors.Wrap(err, "problem with parameters")
@@ -49,14 +74,104 @@ func New(_ context.Context, params ...Parameter) (*Service, error) {
 		log = log.Level(parameters.logLevel)
 	}
 
+	spec, err := parameters.specProvider.Spec(ctx)
+	if err != nil {
+		return nil, errors.Wrap(err, "failed to obtain spec")
+	}
+
+	tmp, exists := spec["SLOTS_PER_EPOCH"]
+	if !exists {
+		return nil, errors.New("failed to obtain SLOTS_PER_EPOCH")
+	}
+	slotsPerEpoch, ok := tmp.(uint64)
+	if !ok {
+		return nil, errors.New("SLOTS_PER_EPOCH of unexpected type")
+	}
+
+	tmp, exists = spec["TIMELY_SOURCE_WEIGHT"]
+	if !exists {
+		// Set a default value based on the Altair spec.
+		tmp = uint64(14)
+	}
+	timelySourceWeight, ok := tmp.(uint64)
+	if !ok {
+		return nil, errors.New("TIMELY_SOURCE_WEIGHT of unexpected type")
+	}
+
+	tmp, exists = spec["TIMELY_TARGET_WEIGHT"]
+	if !exists {
+		// Set a default value based on the Altair spec.
+		tmp = uint64(26)
+	}
+	timelyTargetWeight, ok := tmp.(uint64)
+	if !ok {
+		return nil, errors.New("TIMELY_TARGET_WEIGHT of unexpected type")
+	}
+
+	tmp, exists = spec["TIMELY_HEAD_WEIGHT"]
+	if !exists {
+		// Set a default value based on the Altair spec.
+		tmp = uint64(14)
+	}
+	timelyHeadWeight, ok := tmp.(uint64)
+	if !ok {
+		return nil, errors.New("TIMELY_HEAD_WEIGHT of unexpected type")
+	}
+
+	tmp, exists = spec["SYNC_REWARD_WEIGHT"]
+	if !exists {
+		// Set a default value based on the Altair spec.
+		tmp = uint64(2)
+	}
+	syncRewardWeight, ok := tmp.(uint64)
+	if !ok {
+		return nil, errors.New("SYNC_REWARD_WEIGHT of unexpected type")
+	}
+
+	tmp, exists = spec["PROPOSER_WEIGHT"]
+	if !exists {
+		// Set a default value based on the Altair spec.
+		tmp = uint64(8)
+	}
+	proposerWeight, ok := tmp.(uint64)
+	if !ok {
+		return nil, errors.New("PROPOSER_WEIGHT of unexpected type")
+	}
+
+	tmp, exists = spec["WEIGHT_DENOMINATOR"]
+	if !exists {
+		// Set a default value based on the Altair spec.
+		tmp = uint64(64)
+	}
+	weightDenominator, ok := tmp.(uint64)
+	if !ok {
+		return nil, errors.New("WEIGHT_DENOMINATOR of unexpected type")
+	}
+
 	s := &Service{
 		processConcurrency:           parameters.processConcurrency,
+		chainTime:                    parameters.chainTime,
 		beaconBlockProposalProviders: parameters.beaconBlockProposalProviders,
 		signedBeaconBlockProvider:    parameters.signedBeaconBlockProvider,
 		timeout:                      parameters.timeout,
 		clientMonitor:                parameters.clientMonitor,
+		slotsPerEpoch:                slotsPerEpoch,
+		timelySourceWeight:           timelySourceWeight,
+		timelyTargetWeight:           timelyTargetWeight,
+		timelyHeadWeight:             timelyHeadWeight,
+		syncRewardWeight:             syncRewardWeight,
+		proposerWeight:               proposerWeight,
+		weightDenominator:            weightDenominator,
+		priorBlocks:                  make(map[phase0.Root]*priorBlockVotes),
 	}
 	log.Trace().Int64("process_concurrency", s.processConcurrency).Msg("Set process concurrency")
+
+	// Subscribe to head events.  This allows us to go early for attestations if a block arrives, as well as
+	// re-request duties if there is a change in beacon block.
+	// This also allows us to re-request duties if the dependent roots change.
+	if err := parameters.eventsProvider.Events(ctx, []string{"head"}, s.HandleHeadEvent); err != nil {
+		return nil, errors.Wrap(err, "failed to add head event handler")
+	}
 
 	return s, nil
 }